Fix order number variable size
A local variable in gpu::Scheduler wasn't updated to use 64 bit order
numbers. The value could be truncated as a result.

Fixed: 521959419
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7925952
Commit-Queue: Kyle Charbonneau <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1645528}

+++ gpu/command_buffer/service/scheduler.cc
@@ -498,7 +498,7 @@ Scheduler::Sequence* Scheduler::FindNextTaskFromRoot(Sequence* root_sequence) {
   // First, recurse into any dependency that needs to run before the first
   // task in |root_sequence|. The dependencies are sorted by their order num
   // (because of WaitFence ordering).
-  const uint32_t first_task_order_num = root_sequence->tasks_.front().order_num;
+  const uint64_t first_task_order_num = root_sequence->tasks_.front().order_num;
   DVLOG(10) << "Sequence " << root_sequence->sequence_id()
             << " (order_num: " << first_task_order_num << ") has "
             << root_sequence->wait_fences_.size() << " waits.";
